# PERF: Micro optimize np.linspace