Optimize `np.linspace`:

* Avoid calculation of `_nx.issubdtype` if no `dtype` is specified
* Avoid calling `np.any` in the scalar case

Benchmark with `%timeit linspace(0, 10, 2)` and `%timeit np.linspace([1,2], 10, 5)`

Main:
```
17.1 µs ± 517 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
28.7 µs ± 548 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

PR:
```
8.73 µs ± 153 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
18.5 µs ± 265 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

**Details**

* If no `dtype` is specified, the value for `dtype` is set to `dt` and determined from the input types and `float`. The `float` argument implies the resulting type can never be of integer type, so we skip the call to `_nx.issubdtyp(dtype, _nx.integer)`
* If `_mult_inplace` is `True`, then `delta` is a scalar. We then skip the conversion to a numpy array with `_nx.asarray`
* `np.result_type(..., float)` is almost equivalent to `np.result_type(..., float(num))` (since `num` is of integer type), the former is slightly faster. But for some platforms `test_denormal_numbers` from `test_function_base` fails with this optimization, leaving this out for now. 

<!--         ----------------------------------------------------------------
                MAKE SURE YOUR PR GETS THE ATTENTION IT DESERVES!
                ----------------------------------------------------------------

*  FORMAT IT RIGHT:
      https://www.numpy.org/devdocs/dev/development_workflow.html#writing-the-commit-message

*  IF IT'S A NEW FEATURE OR API CHANGE, TEST THE WATERS:
      https://www.numpy.org/devdocs/dev/development_workflow.html#get-the-mailing-list-s-opinion

*  HIT ALL THE GUIDELINES:
      https://numpy.org/devdocs/dev/index.html#guidelines

*  WHAT TO DO IF WE HAVEN'T GOTTEN BACK TO YOU:
      https://www.numpy.org/devdocs/dev/development_workflow.html#getting-your-pr-reviewed
-->