@eendebakpt 2022-06-25
> Any tradeoffs here? For example, does this optimize one code path at the cost of another?
> 
> A brief comment on that may help make it easier to judge, since I suspect that's a frequent concern with micro-optimizations.

I added some more details to the first comment. But this should make the common paths faster, while keeping all others equally fast. @tylerjereddy