@eendebakpt 2022-06-27
> It seems like the whole point is that `issubdtype` is slow. I am not sure I am a bit fan of it anyway, but I guess fixing _that_ is a different issue and in this case it won't really do to use `dtype.kind in "ui"`...
> 
> Since it is a pretty big difference, I am happy with this. Would be great to proof it with a benchmark though.

`_nx.issubdtype` not that slow:
```
%timeit _nx.issubdtype(float, _nx.integer)
505 ns ± 13.6 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```
But eliminating the 0.5 ms for the common cases is worth it I think. The main gain is on the scalar path and replacing `_nx.any` with `().any()`.

I could not find any benchmark of linspace in the numpy benchmarks, so I added a small benchmark for the scalar and array case.