@mattip 2022-07-07
Since you added a benchmark, could you show the before/after when running `runtests --benchmark` ?