@eendebakpt 2022-07-07
> Since you added a benchmark, could you show the before/after when running `runtests --benchmark` ?

@mattip Here are the benchmark results. The `linspace` tests have improved. There is quite some noise on some of the other tests, I believe this is because my system is not stable enough.

<details>
<summary>python runtests.py --bench-compare main bench_function_base</summary>

```
      before           after         ratio
     [89b942e4]       [3a573b5c]
     <main>           <optimize_linspace>
+     9.80±0.03μs       11.5±0.6μs     1.17  bench_function_base.Sort.time_sort('merge', 'int64', ('uniform',))
+     10.8±0.07μs       12.4±0.3μs     1.15  bench_function_base.Sort.time_argsort('merge', 'int32', ('uniform',))
+       125±0.8μs          141±1μs     1.13  bench_function_base.Sort.time_argsort('merge', 'int32', ('sorted_block', 10))
+       135±0.9μs        149±0.8μs     1.10  bench_function_base.Sort.time_argsort('merge', 'uint32', ('sorted_block', 10))
+       243±0.8μs          264±2μs     1.08  bench_function_base.Sort.time_sort('quick', 'uint32', ('sorted_block', 1000))
+     15.6±0.07μs       16.8±0.3μs     1.07  bench_function_base.Sort.time_argsort('merge', 'float64', ('ordered',))
+      15.7±0.1μs       16.7±0.4μs     1.06  bench_function_base.Sort.time_argsort('merge', 'float32', ('uniform',))
+      11.1±0.1μs      11.7±0.06μs     1.06  bench_function_base.Sort.time_argsort('merge', 'uint32', ('uniform',))
+        52.8±1μs       55.7±0.2μs     1.06  bench_function_base.Sort.time_sort('quick', 'int16', ('ordered',))
+     14.1±0.03μs       14.8±0.2μs     1.05  bench_function_base.Sort.time_argsort('merge', 'uint32', ('reversed',))
-        41.2±2μs       39.2±0.3μs     0.95  bench_function_base.Sort.time_sort('merge', 'int16', ('random',))
-     7.51±0.05μs      7.15±0.04μs     0.95  bench_function_base.Where.time_all_zeros
-         339±1μs        322±0.6μs     0.95  bench_function_base.Sort.time_sort('quick', 'int32', ('sorted_block', 100))
-       109±0.9μs          103±1μs     0.95  bench_function_base.Sort.time_argsort('merge', 'float32', ('sorted_block', 100))
-      8.28±0.2μs       7.76±0.2μs     0.94  bench_function_base.Mean.time_mean_axis(10)
-         261±1μs        244±0.4μs     0.94  bench_function_base.Sort.time_sort('quick', 'int32', ('sorted_block', 1000))
-         321±2μs          299±1μs     0.93  bench_function_base.Sort.time_argsort('quick', 'int32', ('sorted_block', 1000))
-         318±2μs          295±3μs     0.93  bench_function_base.Sort.time_argsort('quick', 'int16', ('sorted_block', 1000))
-         317±1μs          294±1μs     0.93  bench_function_base.Sort.time_argsort('quick', 'uint32', ('sorted_block', 1000))
-       224±0.9μs        203±0.9μs     0.91  bench_function_base.Histogram1D.time_small_coverage
-     12.5±0.06μs       11.2±0.2μs     0.89  bench_function_base.Sort.time_sort('merge', 'uint32', ('reversed',))
-      31.7±0.6μs       24.3±0.2μs     0.77  bench_function_base.Linspace.time_linspace_array
-      20.8±0.7μs       10.5±0.2μs     0.50  bench_function_base.Linspace.time_linspace_scalar

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
```
</details>