@mattip 2022-07-08
The speedup (the results are presented to appear to be a slowdown) is significant, it seems this optimization is worthwhile.